requirements.txt
azure-core
azure-ai-inference
python-dotenv>=0.19.0
python-telegram-bot